Move component to its own file
