+++ src/NavBar/SmallNavBar.tsx
@@ -1,0 +1,128 @@
+// @ts-nocheck
+import React from "react";
+import PropTypes from "prop-types";
+import { Branding } from "../Branding";
+import { Flex } from "../Flex";
+import theme from "../theme";
+import NavBarSearch from "../NavBarSearch/NavBarSearch";
+import { PreventBodyElementScrolling } from "../utils";
+import MobileMenu from "./MobileMenu";
+import {
+  BrandingLink,
+  MenuIcon,
+  MobileMenuTrigger,
+  NavBarBackground,
+  SmallHeader,
+} from "./NavBar";
+
+class SmallNavBarNoState extends React.Component {
+  constructor() {
+    super();
+    this.navRef = React.createRef();
+  }
+
+  componentDidUpdate(prevProps) {
+    if (this.props.menuState.isOpen && !prevProps.menuState.isOpen)
+      this.navRef.current.scrollTop = 0;
+  }
+
+  isSmallScreen() {
+    const { breakpointLower, width } = this.props;
+
+    return width < pixelDigitsFrom(breakpointLower);
+  }
+
+  render() {
+    const {
+      menuData,
+      menuState: { isOpen, toggleMenu, closeMenu },
+      subtext,
+      brandingLinkHref,
+      brandingLinkTo,
+      themeColor,
+      ...props
+    } = this.props;
+    return (
+      <SmallHeader ref={this.navRef} isOpen={isOpen} {...props}>
+        <NavBarBackground
+          backgroundColor={getThemeColor(themeColor).background}
+        >
+          <BrandingLink
+            aria-label="Nulogy logo"
+            display="block"
+            height={subtext && !this.isSmallScreen() ? "56px" : "40px"}
+            my={subtext && !this.isSmallScreen() ? "-8px" : null}
+            underline={false}
+            href={brandingLinkHref}
+            to={brandingLinkTo}
+          >
+            <Branding
+              logoColor={getThemeColor(themeColor).logoColor}
+              logoType={this.isSmallScreen() ? "lettermark" : "wordmark"}
+              subtext={this.isSmallScreen() ? null : subtext}
+            />
+          </BrandingLink>
+          <Flex
+            justifyContent="flex-end"
+            style={{ flexGrow: 1, margin: `0 0 0 ${theme.space.x3}` }}
+          >
+            {menuData.search && (
+              <Flex maxWidth="18em" alignItems="center" px="0">
+                <NavBarSearch {...menuData.search} />
+              </Flex>
+            )}
+            {(menuData.primaryMenu || menuData.secondaryMenu) && (
+              <MobileMenuTrigger
+                {...getThemeColor(themeColor)}
+                onClick={toggleMenu}
+                aria-expanded={isOpen ? true : null}
+              >
+                <MenuIcon isOpen={isOpen} />
+              </MobileMenuTrigger>
+            )}
+          </Flex>
+        </NavBarBackground>
+        {isOpen && (
+          <PreventBodyElementScrolling>
+            <MobileMenu
+              themeColorObject={getThemeColor(themeColor)}
+              subtext={subtext}
+              includeSubtext={this.isSmallScreen()}
+              menuData={menuData}
+              closeMenu={closeMenu}
+            />
+          </PreventBodyElementScrolling>
+        )}
+      </SmallHeader>
+    );
+  }
+}
+
+SmallNavBarNoState.propTypes = {
+  menuState: PropTypes.shape({
+    isOpen: PropTypes.bool,
+    toggleMenu: PropTypes.func,
+    closeMenu: PropTypes.func,
+  }).isRequired,
+  menuData: PropTypes.shape(MenuDataPropTypes),
+  subtext: PropTypes.string,
+  brandingLinkHref: PropTypes.string,
+  brandingLinkTo: PropTypes.string,
+  breakpointLower: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+  width: PropTypes.number,
+  themeColor: PropTypes.oneOf(["blue", "white"]),
+};
+
+SmallNavBarNoState.defaultProps = {
+  menuData: null,
+  subtext: null,
+  brandingLinkHref: "/",
+  brandingLinkTo: undefined,
+  breakpointLower: theme.breakpoints.small,
+  width: undefined,
+  themeColor: undefined,
+};
+
+const SmallNavBar = withMenuState(SmallNavBarNoState);
+
+export default SmallNavBar;
